fish: make `fn-clear` ignore fifs functions

+++ fish/fun/special.fish
@@ -102,6 +102,7 @@ function fn-clear
     for file in ~/.config/fish/functions/*.fish
         set function_name (basename $file '.fish')
         if not contains $function_name $list
+            and not string match -qr '^_?fifc' $function_name
             rm $file
             echo 'cleared: '$function_name
         end
